Houskeeping: deleting or moving files
