--- analysis/__init__.py
@@ -1,1 +1,0 @@
-"""Analysis helpers for Parquet outputs and summaries."""
