+++ utils/__init__.py
@@ -1,0 +1,4 @@
+"""Utility helpers for analysis and plotting.
+
+This package consolidates reusable notebook utilities.
+"""
